Better float/double tests

+++ tests/round-trip-doubles/source.cpp
@@ -6,7 +6,7 @@ int main() {
 #ifndef cout
     std::cout << std::setprecision(16);
 #endif
-    float x;
+    double x;
     while (std::cin >> x) {
         std::cout << x << std::endl;
     }
